refactor: auth 도메인의 잘못 배치된 dto 파일을 위치 수정
