+++ backend/src/main/java/com/icehufs/icebreaker/domain/auth/dto/object/JwtToken.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.domain.auth.domain.vo;
+package com.icehufs.icebreaker.domain.auth.dto.object;
 
 public record JwtToken(
 	String accessToken,
