+++ backend/src/main/java/com/icehufs/icebreaker/domain/auth/service/implement/AuthServiceImplement.java
@@ -9,7 +9,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.icehufs.icebreaker.domain.auth.domain.vo.JwtToken;
+import com.icehufs.icebreaker.domain.auth.dto.object.JwtToken;
 import com.icehufs.icebreaker.domain.auth.dto.request.RegenerateTokenRequestDto;
 import com.icehufs.icebreaker.domain.auth.dto.response.LogoutResponseDto;
 import com.icehufs.icebreaker.domain.auth.dto.response.RegenerateTokenResponseDto;
